Moved some code from the evaluation module to the new value_types module
